requirements.txt
kafka-python~=2.2.13
pandas~=2.3.0
s3fs